+++ packages/page-spy-base/src/client.ts
