+++ packages/page-spy-base/src/config.ts
