+++ packages/page-spy-base/src/constants.ts
