+++ packages/page-spy-base/src/request-item.ts
@@ -1,6 +1,6 @@
 import { SpyNetwork } from '@huolala-tech/page-spy-types';
 
-export default class RequestItem implements SpyNetwork.RequestInfo {
+export class RequestItem implements SpyNetwork.RequestInfo {
   id = '';
 
   name: string = '';
