+++ packages/page-spy-base/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "target": "ESNext",
+    "outDir": "dist"
+  },
+  "include": ["src"]
+}
