+++ packages/page-spy-browser/package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "@babel/runtime": "^7.13.0",
     "@huolala-tech/page-spy-types": "^1.9.0",
+    "@huolala-tech/page-spy-base": "*",
     "copy-to-clipboard": "^3.3.1"
   },
   "devDependencies": {
