+++ packages/page-spy-mp-base/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@huolala-tech/page-spy-mp-base",
+  "version": "1.9.0",
+  "description": "The base of PageSpy for miniprogram.",
+  "license": "MIT",
+  "main": "dist/iife/index.min.js",
+  "module": "dist/esm/index.min.js",
+  "types": "types/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "dependencies": {
+    "@huolala-tech/eval5": "^0.0.1",
+    "@huolala-tech/page-spy-base": "*",
+    "@huolala-tech/page-spy-types": "*"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "homepage": "https://www.pagespy.org/",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/HuolalaTech/page-spy.git"
+  },
+  "bugs": {
+    "url": "https://github.com/HuolalaTech/page-spy/issues"
+  },
+  "files": [
+    "dist",
+    "types"
+  ],
+  "keywords": [
+    "pagespy",
+    "debug",
+    "remote",
+    "inspect",
+    "devtools"
+  ]
+}
