+++ packages/page-spy-mp-base/src/config.ts
