+++ packages/page-spy-mp-base/src/plugins/console.ts
