+++ packages/page-spy-mp-base/src/plugins/error.ts
