+++ packages/page-spy-mp-base/src/plugins/network/proxy/base.ts
