+++ packages/page-spy-mp-base/tests/mock/storage.ts
