+++ packages/page-spy-mp-base/tests/plugins/console.test.ts
